Use obs_frontend_add_dock_by_id
The API `obs_frontend_add_dock` was deprecated so that use the new API.

Added a workaround to show the dock at the initial creation.

Due to the limitation of the new API, when the dock is closed by X
button, the dock won't be deleted. To delete the dock, user need to use
`Close` in the context-mene.